requirements.txt
pandas
numpy
matplotlib
tiktoken
watchdog          # <─ new
python-dotenv
